move variables out of analysis and into its own module
