+++ bids/analysis/__init__.py
@@ -1,4 +1,4 @@
-from .variables import load_variables
+from bids.variables import load_variables
 from .analysis import Analysis
 
 __all__ = ['load_variables', 'Analysis']
