+++ bids/variables/collections.py
