+++ bids/variables/entities.py
